fix: modify how cd /vagrant is done at SSH to show motd again
Before this commit, the `cd /vagrant` at SSH was done with the following line in the Vagrantfile:
`config.ssh.extra_args = ["-t", "cd /vagrant; bash --login"]`

This had the problem that executing `bash --login` didn't show the `motd`
(message of the day) at login, so the `Vagrant VM` welcome banner and the
tmux command were not shown.

This also had another problem: it broke the ability to use the
`vagrant ssh -- <command>` option.